% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{bestParameters}
\alias{bestParameters}
\title{bestParameters}
\usage{
bestParameters(v, nbK = 10)
}
\arguments{
\item{v}{the estimated variances of the diff k operator}

\item{nbK}{number of diff k elements to consider}
}
\value{
a list with an estimation of the best parameters for Eta2, Nu2 and phi
}
\description{
iteration of the least square criterion for a grid of the phi parameter
}
