% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{cost}
\alias{cost}
\title{cost}
\usage{
cost(v, sdEta2, sdNu2, phi, nbK = 10)
}
\arguments{
\item{v}{the estimated variances of the diff k operator}

\item{sdEta2}{the Random Walk variance}

\item{sdNu2}{the AR(1) variance}

\item{phi}{the autocorrelative AR(1) parameter}

\item{nbK}{number of diff k elements to consider}
}
\value{
the value of the sum of squares
}
\description{
the least-square value
}
