% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGenerator.R
\name{dataARRW}
\alias{dataARRW}
\title{dataARRW}
\usage{
dataARRW(n = 1000, sdEta2 = 1, sdNu2 = 1, phi = 0.9,
  poisParam = 0, meanGap = 2)
}
\arguments{
\item{n}{number of data points}

\item{sdEta2}{variance in Random Walk}

\item{sdNu2}{variance in AR(1)}

\item{phi}{AR(1) autocorrelation parameter}

\item{poisParam}{a probability in Poisson distribution for changepoint density}

\item{meanGap}{mean gap of the jumps}
}
\value{
a vector of simulated data with y = the data, mu = the signal, changepoints = changepoint indices
}
\description{
Generate data with the ARRW model
}
\examples{
myData <- dataARRW(100)
}
