% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGenerator.R
\name{dataRWAR}
\alias{dataRWAR}
\title{Generating data function with the deCAFS model}
\usage{
dataRWAR(
  N = 1000,
  sdEta = 1,
  sdNu = 1,
  phi = 0.5,
  type = c("none", "up", "updown", "rand1"),
  nbSeg = 20,
  jumpSize = 1,
  seed = 42
)
}
\arguments{
\item{N}{number of data points}

\item{sdEta}{standard deviation in Random Walk}

\item{sdNu}{standard deviation in AR(1)}

\item{phi}{AR(1) autocorrelation parameter}

\item{type}{possible scenarios for the jump structure}

\item{nbSeg}{number of segments}

\item{jumpSize}{max size of the jumps}

\item{seed}{random number generator seed (default is 42)}
}
\value{
A list containing:
\describe{
\item{\code{y}}{the data sequence,}
\item{\code{signal}}{the underlying signal without the superimposed AR(1) noise,}
\item{\code{changepoints}}{the changepoint indices}
}
}
\description{
Generate a Realization from the deCAFS model (see references for further details).
\deqn{y_t = \mu_t + \epsilon_t}
where
\deqn{\mu_t = \mu_{t-1} + \eta_t + \delta_t, \quad \eta_t \sim N(0, \sigma_\eta^2), \ \delta_t \ \in R}
and
\deqn{\epsilon_t = \phi \epsilon_{t-1} + \nu_t \quad \nu_t \sim N(0, \sigma_\nu^2)}
}
\examples{
myData <- dataRWAR(1000, sdEta = 0.1, sdNu = 0.1, type = "rand1",  nbSeg = 10, seed = 86)
}
