% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{estimVar}
\alias{estimVar}
\title{estimVar}
\usage{
estimVar(y, nbK = 10)
}
\arguments{
\item{y}{An object created by the dataARRW function}

\item{nbK}{number of diff k elements to consider}
}
\value{
a list with a varEst element (= a vector)
}
\description{
Estimation of the variances for the diff k operator k = 1 to nbK
}
