% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{evalEtaNu}
\alias{evalEtaNu}
\title{evalEtaNu}
\usage{
evalEtaNu(v, phi, nbK = 10)
}
\arguments{
\item{v}{the estimated variances of the diff k operator}

\item{phi}{the autocorrelative AR(1) parameter}

\item{nbK}{number of diff k elements to consider}
}
\value{
a list with an estimation of the variances Eta2 and Nu2
}
\description{
Evaluation of the variances Eta2 and Nu2
}
