% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2Threshold.R
\name{l2Threshold}
\alias{l2Threshold}
\title{l2Threshold}
\usage{
l2Threshold(data, beta, lambda)
}
\arguments{
\item{data}{a time-series}

\item{beta}{the penalty value}

\item{lambda}{1/variance2}
}
\value{
a list of changepoints with their jump sizes
}
\description{
Simple linear threshold changepoint detection method
}
