% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{one.simu}
\alias{one.simu}
\title{one.simu}
\usage{
one.simu(i, n = 10^5, sdEta2 = 1, sdNu2 = 1, phi = 0.5,
  poisParam = 0, meanGap = 2, nbK)
}
\arguments{
\item{i}{index of the simulation}

\item{n}{number of data points}

\item{sdEta2}{variance in Random Walk}

\item{sdNu2}{variance in AR(1)}

\item{phi}{AR(1) autocorrelation parameter}

\item{poisParam}{a probability in Poisson distribution for changepoint density}

\item{meanGap}{mean gap of the jumps}

\item{nbK}{number of diff k elements to consider}
}
\value{
a dataframe with the inital parameters and the estimated ones
}
\description{
Perform the analysis form data generation to parameter estimation
}
