% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotARRWdiff}
\alias{plotARRWdiff}
\title{plotARRWdiff}
\usage{
plotARRWdiff(y)
}
\arguments{
\item{y}{An object created by the dataARRW function}
}
\description{
plot y_{t+1} - y_t signal and changepoint locations to identify obvious changepoints
}
