% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotRWARdiff}
\alias{plotRWARdiff}
\title{plotRWARdiff}
\usage{
plotRWARdiff(y)
}
\arguments{
\item{y}{An object created by the dataARRW function}
}
\description{
plot y_{t+1} - y_t signal and changepoint locations to identify obvious changepoints
}
