% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotVarVarEstim}
\alias{plotVarVarEstim}
\title{plotVarVarEstim}
\usage{
plotVarVarEstim(v, sdEta2, sdNu2, phi, nbK = 10)
}
\arguments{
\item{v}{the estimated variances of the diff k operator}

\item{sdEta2}{the Random Walk variance}

\item{sdNu2}{the AR(1) variance}

\item{phi}{the autocorrelative AR(1) parameter}

\item{nbK}{number of diff k elements to consider}
}
\description{
plot the estimated variances v_k against the true variances for the diff k operator (y_{t+k} - y_t) for k = 1 to nbK
}
