% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGenerator.R
\name{scenarioGenerator}
\alias{scenarioGenerator}
\title{Signal scenario}
\usage{
scenarioGenerator(
  N,
  type = c("none", "up", "updown", "rand1"),
  nbSeg = 20,
  jumpSize = 1,
  seed = 42
)
}
\arguments{
\item{N}{number of data points}

\item{type}{possible scenarios for the jump structure}

\item{nbSeg}{number of segments}

\item{jumpSize}{max size of the jumps}

\item{seed}{random number generator seed (default is 42)}
}
\description{
Generate a signal scenario
}
