% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{evalEtaNu}
\alias{evalEtaNu}
\title{RW and AR(1) variance estimations with fixed AR(1) parameter}
\usage{
evalEtaNu(v, phi)
}
\arguments{
\item{v}{the estimated variances of the diff k operator}

\item{phi}{the autocorrelative AR(1) parameter}
}
\value{
a list with an estimation of the variances Eta2 and Nu2
}
\description{
Evaluation of the variances Eta2 and Nu2
}
